refactor: try to improve naming, seperate schemas
