+++ src/containers/ImageEditor/ImageEditor.jsx
@@ -14,7 +14,7 @@ import defined from 'defined';
 import Types from 'slate-prop-types';
 import { Crop, FocalPoint } from '@ndla/icons/editor';
 import { injectT } from '@ndla/i18n';
-import { getSchemaEmbed } from '../../components/SlateEditor/schema';
+import { getSchemaEmbed } from '../../components/SlateEditor/editorSchema';
 import { EmbedShape, EditorShape } from '../../shapes';
 import ImageTransformEditor from './ImageTransformEditor';
 import ImageAlignButton from './ImageAlignButton';
